Update elasticsearch.rb
I find in a cluster with many indexes and some are huge, the current filter, without the capability to pin point the specific index, will cause huge load on server because it is trying to search all indexes.

the proposed change is to add an optional index parameter, if not specified, working as previously (default to _all), and if specified, will limit to search for that index (not sure if wildcard works, did not test yet). This will make query on a specific index only to reduce CPU load.