Fix formatting issues
Not sure it will fix it but let's give it a try. 

The following section:

> The example below reproduces the above example but utilises the query_template.  This query_template represents a full Elasticsearch query DSL and supports the standard Logstash field substitution syntax.  The example below issues the same query as the first example but uses the template shown.

Is appearing as a code:

```
The example below reproduces the above example but utilises the query_template.  This query_template represents a full
Elasticsearch query DSL and supports the standard Logstash field substitution syntax.  The example below issues
the same query as the first example but uses the template shown.
```

This commit tries to fix it. It also corrects the indentation of the code examples.