Document the new behaviour...
For it to work as discussed, we need to remove the default value for `ssl`.
This is not a breaking change. Prior to this, `ssl => false` triggered no code
path and didn't fail if https hosts were specified with `ssl => false`.
Therefore users were able to leave `ssl` unspecified and pass a bunch of
https hosts.

Now that we're being more strict whenever `ssl` is specified explicitly,
this prior way of specifying https in hosts didn't work anymore.
By removing the default value, we can have a ternary value:

* Not specifying it will continue to allow people to use https schemes in hosts
* Specifying `ssl`  will now only accept schema-less hosts or the proper
kind of schema (e.g. https if ssl true). It will continue to raise an
error if ssl is true and an http schema is specified; and it will now
raise also if ssl is false, and an https schema is specified.