normalise hosts to `URI::Generic`s, which reliably preserve defaults
An upstream bug in the Elasticsearch Ruby Client's handling of `String` host
arguments that begin with a schema (e.g., `https://localhost`) causes it to
default to port 80 or 443, depending on the schema, instead of Elasticsearch's
port 9200.

Since the Elasticsearch Ruby Client will accept a `URI` in this case, and will
correctly handle falling through to appropriate defaults, we normalise to
`URI::Generic`, which does not have a default port.

We absorb the `ssl => true` case into this normalisation, as its previous
implementation prevented the use of non-default ports in the array provided
to `hosts`.

We also add support for IPv6 addresses, requiring a square-bracketed notation
when used in conjunction with a specified port.
(see: RFC-3986)

Supersedes: https://github.com/logstash-plugins/logstash-filter-elasticsearch/pull/104
Resolves:   https://github.com/logstash-plugins/logstash-filter-elasticsearch/issues/110
Resolves:   https://github.com/logstash-plugins/logstash-filter-elasticsearch/issues/111