avoid holding refs to threads, rely on manticore threadsafety
Since our transport class Manticore is explicitly threadsafe and
by default creates a pool with up to 50 connections, we do not need
an individual instance per worker thread.

Using a shared instance also closes a memory leak that was caused
by using a strong reference to the thread itself as a key in the
routing to per-thread clients, which was not cleaned up on pipeline
failure or completion.