fix: free resources on close
Our transport class Manticore is explicitly threadsafe and by default creates
a pool with up to 50 connections; therefore we do not need an individual
instance of Manticore per worker thread. Using a shared instance resolves a
memory leak that was caused by using a strong reference to the worker threads
themselves as a key in the routing to per-thread clients, which prevented
those threads from being garbage-collected.